requirements.txt
basedosdados==1.6.1
slackclient==2.9.3
python-dotenv==0.19.2